chore(duplication): share vscode tokenized request helper (#5479)
